Use user-defined state types instead of int.
Before, a state was denoted by an `int`. This was flexible and generally
ok for the state machine itself. Because of the implicit conversion, a
user could also use a (non type-safe) enum. However, since the states
were always returned as int, it was not quite safe for the client.

This commit now specifies the type that declares the states. With this,
it is now possible to use type-safe `enum class`es.

For the transition, one can define the state type as int.

Another change that goes together with this is that the 2 pre-defined
states (Initial, Final) are no longer defined. The Initial state must
be defined as a template argument, and the Final state is no longer
specified. The reason for the Final state is that there might be none
or many, and therefore the Final state is not really special. The
`is_final()` method has been removed.

Since the Initial state is now specified as a template argument, the
`init()` method is also no longer required.

Integers can still be used as a type like the following example. Use this
for the transition.

    using F = FSM::Fsm<int, 1>;
    F fsm;
    std::vector<F::Trans> transitions = {...}

And then F can be used instead of the previous definition.

Tests are updated.