Add an optional parameter to specify a state when reset the state machine.
This allows to reset to a specific state, not only the initial state.
This is useful to implement some form of history, although the
history part needs to be implemented externally.

The default behaviour when reset is still to set the state to the
Initial state.

Also add a new test case.